image_server/pyproject.toml
[project]
name = "vista3d-image-server"
version = "0.1.0"
description = "Standalone FastAPI image server for Vista3D"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "nibabel>=5.3.2",
    "numpy>=2.0.0",
]

# no build backend required; runtime-only project


